requirements.txt
alembic==1.0.7
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.14.5
Click==7.0
dominate==2.6.0
Flask==1.0.2
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.1.2
MarkupSafe==1.1.0
numpy==1.20.2
pycparser==2.20
PyJWT==1.4.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2021.1
six==1.12.0
SQLAlchemy==1.3.6
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
